refactor: isolate package map implementation
